backend/requirements.txt
fastapi==0.111.0
uvicorn==0.22.0
pydantic==2.8.2
pydantic-settings==2.2.1
email-validator>=2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.41
psycopg[binary]
supabase==1.0.3
pypdf2==3.0.1
spacy==3.5.3
pytesseract==0.3.10
pytest==7.3.1
httpx==0.24.1
celery==5.2.7
redis==4.5.5
google-generativeai==0.8.3
